Adding Xcode project etc. for new external: jmod.delta. This is a Jamoma port of tl.delta from tl.objects, and external calculating first order differences for floats.
I have successfully compiled and tested development version of the object in Mac OSX Max 4.5.7. I have not yet tested Deployment version in Max 4.6.

Remains to be done: Everything Windows-related + committing compiled Mac object. As my Mac_build folder is a bit screwed up at the moment (in order to get it to work on 4.5.7) I don't want to mess up the build folder for others.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@951 464221e2-0a0f-0410-983b-f8d57d210c17