Added Deployment builds of jmod.delta, jmod.delta2 and jmod.velocity to build_mac.
I have not touched the built versions of the rest of the externals.

WARNING: jmod.velocity is currently crashing Max. There is something wrong in the new function, but I haven't been able to figure out exactly what yet. Sorry about that. I still wanted to commit, in case Tim might have the time to take a look at it.

%-/

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@956 464221e2-0a0f-0410-983b-f8d57d210c17