Adding new module for recording video to QuickTime movie.
This also introduce two components dealing with standard (common) attributes of Jitter Matrix Objects in a class-like way. One component is a bpathcer for use in inspectors in video modules, causing the module to inherit a number of parameters and messages. The other is a routing utility for use inside the algorithm.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@1102 464221e2-0a0f-0410-983b-f8d57d210c17