Adding new external: jmod.cubic_interpolate. This performs cubic interpolation on a stream of floats.
As my own build_mac folder is a mix of new and old versions of objects in order to work on Max 4.5.7/PPC, I'm not committing updated version of it. My PC is still tied up in an installation, so I'm unable to compile for windows right now. I still believe I've modified the Windows project so that it should be simple to compile.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@1137 464221e2-0a0f-0410-983b-f8d57d210c17