Changed the way jmod.colorpicker.mod is working: It is now describing colors as HSL: Hue (0-360), Saturation (0-100) and Lightness (0-100).
This solved a redrawing problem with the color swatches that would make the be attracted toward red, and also seems like a more sensible way of describing and dealing with colors in a text format.

I'm adding a new components containing the colorswatch and number boxes for HSL. I am also for the time being depending on tap.colorswatch as Tim has offered to move that external to Jamoma.

Also updated jmod.palette.mod to reflect the change.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@1146 464221e2-0a0f-0410-983b-f8d57d210c17