Added a qlim on the bang that drives the gui constructor.
In the short term, this is a fix for the problems Trond noted in his last commit with the ambi modules.  

In the long term, this raises a red flag that this bang is potentially being triggered from multiple threads -- and much of the code in the new core isn't really threadsafe at the moment.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@1636 464221e2-0a0f-0410-983b-f8d57d210c17