Windows Installer...   Apparently this was never really set up before.  Dealing with the third-party extern packages is a mess, so we should talk about how we are dealing with this in Albi...
The structure of the installer is that there is a merge module for the Jamoma files that need to be installed, and then an installer that wraps the merge module.  This is done so that the merge module can be used in other installers as well.

A problem is that the various files used in the installer interface must be RTF.  This means I had to make special copies of both the license and readme.  I suggest we change both of these files to RTF files in the repository in the near future.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@1638 464221e2-0a0f-0410-983b-f8d57d210c17