Qmoma - This is the initial check-in of an idea I've been wanting to do for years.  It is basically the cuelist module as a native Mac application.
It uses NSOutlineView for the cue list interface (collapsing cues, etc.).  Ultimately it will send messages to Max (and/or other hosts) on the local (or remote) machine(s).  

Hopefully it will also be able to serve as a high-resolution external time source for future things we might want to do in Jamoma, but this is a future project.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/experimental@2205 464221e2-0a0f-0410-983b-f8d57d210c17