Renamed 'clear' to 'clearMessage' so that the message will be published when wrapped as a Max external.  Need to apply this change to other objects too...
Also got code to compile by using constants just added in my last check-in.

Finally, fixed a bad bug where the vs didn't get reset when moving from processing one channel to the next.  This problem is present in other objects and so needs to be fixed in them too.

git-svn-id: http://ttblue.googlecode.com/svn/branches/active@172 bc2ad3ca-9a4d-0410-8d14-35da9c29b421