More work on basic components to use for spatialisation/SpatDIF:
- jcom.aed2ambimonitor
- jcom-ambimonitor2aed

Updated to work with new coordinate conventions. Same unit lenght is used for distance in spherical coordinates and xyz in Cartesian. Added attribute @radius that can be used to describe the radius of the circle covered by ambimonitor. This way ambimonitor can be used for setups of varying sizes.

- jcom.xyz2ambimonitor
- jcom.ambimonitor2xyz

New components for mapping between icst.ambimonitor and SpatDIF Cartesian coordinates. Also use the @radius.

Tweaks to various help files.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2325 464221e2-0a0f-0410-983b-f8d57d210c17