The humble beginnings of a new time infrastructure for Jamoma.  There is a framework that implements the core time functions, and a couple of externals built against the framework.
jcom.phasor~ is like a phasor~ object but driven by the Jamoma.framework.  jcom.metro is like metro but also driven by the Jamoma.framework.  jcom.time is an object to control the framework itself.  This stuff is all pretty rudimentary, and does not yet entirely work.  However, some aspects are working -- like the synchronization of phasors.  For these initial experiments and implementations, MSP is being used as the master driving clock.  In future implementation we can hopefully have some other options as well - such as our own high-priority thread for generating a master clock.

Once these basic objects are fully working and the system proves itself, then we can start to integrate it with the RampLib and Jamoma's core to have better control over automation of parameters.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/experimental@2364 464221e2-0a0f-0410-983b-f8d57d210c17