hooray -- jcom.send~ essentially works now.  It can certainly be made a lot cooler, but the important part (the basic implementation) is fully functional: multiple jcom.send~ objects can be used to communicate audio signals to modules "wirelessly", and the signals are summed.
My basic test is to open the jmod.degrade~ help patcher and the jcom.send~ help patcher at the same time and play with them.  At least in this limited case, things are working very well.  And in fact I think the jcom.in~ dsp restructuring is more efficient as well.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2461 464221e2-0a0f-0410-983b-f8d57d210c17