Added initial code for getting our own callbacks from CoreAudio to drive the Jamoma clock.  The code is #ifdef'd out until I can test, tweak, and adjust it some more.
At the moment this code will update the Jamoma clock at the start of each vector as provided directly from the audio interface hardware.  Apparently that setting for the default driver on my MBP is 256 samples, which at 44100 Hz is about 5.8 ms.  Theoretically it is easy to change this by changing the I/O Vector setting -- but that's annoying because it has other performance ramifications too.  Maybe I just need to get over this?


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2484 464221e2-0a0f-0410-983b-f8d57d210c17