Initial implementation of hub "internals" -- that is messages and parameters that are embedded inside of the hub.
This means that we now have properly implemented (and documented) messages to the hub.  For example, /preset/interpolate is actually a jcom.message object that then calls a method in the hub.

At the moment we are only doing this for things that were already in the hub, thus we haven't actually lost any pattr functionality just yet.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2791 464221e2-0a0f-0410-983b-f8d57d210c17