The maintenance of this list is kind of annoying, so maybe I should make something better...
Regardless, with the way it is currently set up, any time we add a dataspace name to the list we also need to increase the numDataspaces variable or we will not allocate enough memory for the names.  It looks like a couple of dataspaces have been added recently, so I just upped the numDataspaces variable accordingly.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2816 464221e2-0a0f-0410-983b-f8d57d210c17