Added (untested) objects: * Crossfade * Gain * Limiter (only half done) * Lowpass Butterworth * Overdrive (only a stub)
Also changed the way we handle notification when a parent class's parameter changes, and the subclass needs to know about it.  Demonstrated in the TTDCBlock code (but yet propagated to the other objects).

Added missing utility functions (such as dB conversions).

Created some more methods for registering parameters with only the things that are needed.  This was done because of strange behaviour I was seeing.  I'm no longer convinced that this was the culprit, but they are still handy to use.

Fixed constness of some arguments (such TTDegrade attr setters).

At the moment things are crashing badly, and there looks to be memory corruption.  Why?  One thing I've seen is that the tt.degrade~ example tries to set the global sr parameter before that parameter exists.  While that isn't good, it shouldn't cause this crashing.  


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@236 bc2ad3ca-9a4d-0410-8d14-35da9c29b421