Hopefully this check-in doesn't break anything on the Mac...
The DLL is now legit (though still missing some exports).  I've tested it with a local copy of jcom.change, though I won't commit that jcom.change until the DLL is finished.

A lot of this is just making functions extern "C" for exporting the symbols.  Also created a .def file for the DLL.  It's not fleshed out yet, but has enough in it to compile jcom.change.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2879 464221e2-0a0f-0410-983b-f8d57d210c17