In the process of implementing ramplib the symbol ps_ramp had been changed from gensym("ramp") to gensym("ramp/drive").
When jcom.parameter receives a list, it looks for the 2nd last member to see if it equals ps_ramp. So when ps_ramp was changed, this test broke.

I've introduced a new symbol ps_ramp_drive = gensym("ramp/drive") and restored ps_ramp to gensym("ramp").

Ramps now seem to work again.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2910 464221e2-0a0f-0410-983b-f8d57d210c17