Removed a non-functional URL.  Also removed the Copperlan item, since no information about it is yet publicly available.
Looking at the remainder of the list, it seems like many of these are irrelevant for the paper though.  For example: yes, SuperCollider uses OSC, but what does that have to do with standardizing an OSC namespace.  Maybe there is something, but then we will need to know what it is.  I guess the point is we need to make an argument for each point here that it has something to do with *standardizing* OSC namespaces.  Yes?


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@2943 464221e2-0a0f-0410-983b-f8d57d210c17