New TTThread class for dealing with threading in TTBlue.  It doesn't do very much interesting yet.
The impetus to create it now was to create mechanism for making a thread sleep.  The idea was that we could sleep the main thread in Max very briefly before deleting an old FunctionUnit, and this would allow things to clear up if anything was still sitting on the scheduler thread.  Unfortunately, it doesn't appear that we can sleep the thread, and thus this strategy doesn't work.  Never-the-less, this will be useful in the future in other contexts.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@330 bc2ad3ca-9a4d-0410-8d14-35da9c29b421