Dave's commit drew my attention to this code.
I'm not sure if we can rely on Max passing us the right amount of memory or not.  It is safe to assume 1 atom, but I'm not sure about more than that and need to research it.  

Or we could just use the sysmem_ptrsize() (or whatever it is called) function.  Maybe that would be better anyway...


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@3095 464221e2-0a0f-0410-983b-f8d57d210c17