Warning: This check-in breaks a lot of stuff!  Don't update unless you are ready to do some work fixing things...
One of the goals of Tap.Tools 3, and TTBlue, is to move toward working with a 64-bit audio chain.  To that end all audio signals and vectors in TTBlue are now 64-bit.  In many cases the internal workings of TTBlue objects were already using 64-bit floating-point precision.  Where they weren't they now are.

One very notable problem with this is that Max still uses 32-bit audio vectors.  To handle this I've expanded the way the setVector() method works so that it can be set as before with a 64-bit vector.  If a 32-bit vector is passed in then we allocate a 64-bit buffer and copy the 32-bit values into the new buffer.  This will be fine for getting audio into our 64-bit chain, but all Max objects that process audio are going to need to be adapted for this, and also they will need to copy the 64-bit audio back to their 32-bit vector afterwards.

All of this copying will probably be slow.  How slow?  I don't know.  Maybe it won't be a problem.  Maybe it will.  I guess we need to find that out.

A few minor things in TTAudioSignal also changed to make direct member access forbidden and thus force the use of accessors.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@343 bc2ad3ca-9a4d-0410-8d14-35da9c29b421