Adding the resonant 2-pole and 4-pole filters.
This is one of the things that was in a commit that never made it through from my Windows machine when it gave me a Blue Screen of Death.  The click-of-death from the hard drive started about a week ago and now all is lost.  I've now purchased a new hard drive and am waiting for it to format so I can install Windows afresh.  Goody goody.

git-svn-id: http://ttblue.googlecode.com/svn/branches/active@372 bc2ad3ca-9a4d-0410-8d14-35da9c29b421