I discovered more ESP problems on Windows (via crashes in some other Tap.Tools objects).  While we had earlier dealt with the problem in attributes, we had not fully dealt with the problem in messages.
Since I found the code required to patch this for messages somewhat offensive (to call a clear() method you would have to actually create 2 dummy objects and then pass them to clear(foo, bar) etc.), I decided to plow ahead with one of the changes proposed for the meeting in Genoa: setting flags for messages and attributes.  Actually, I didn't do the attributes yet (though I did create an enum of flags for it).

So now we can create a message and set the kTTMessagePassNone flag and then call methods like clear() with no args.  This required the minor re-writting of messages to be C++ objects like what we already do with attributes.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@457 bc2ad3ca-9a4d-0410-8d14-35da9c29b421