A partial fix to Pascal's report for numbers not working as arguments to jcom.oscroute.
However, the real problem is that Max 5 isn't behaving the same as it did in Max 4 in terms of using quotes to force a number to be interpreted as symbol.  For example, type:
	sel ".1"
into an object box and see what happens.  I'm taking up a discussion about this behavior with by C74 colleagues...


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@3730 464221e2-0a0f-0410-983b-f8d57d210c17