Reverting SVN Rev 507, and instead fixing Issue 14 in a different way:
Now the clear() method for TTValue sets the number of elements to zero so that we can build an array up dynamically with a the append() method.

The = operators then all set the number of values to 1, which they should probably do anyway.  If this breaks anything, it will be because there is an array and we used as shortcut (=) to set the first item of the array (and now the array will be reduced to a size of one with =).


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@575 bc2ad3ca-9a4d-0410-8d14-35da9c29b421