More work on new external: jcom.oscinstance.
Cleaning up. Also now the instance is output form the 2nd outlet, with one caveat: It's currently always of type symbol. Might need to consult gurus from Kansas to figure out how to solve this.

For the time being, tap.typecheck is introduced to the help file while solving this issue.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@3873 464221e2-0a0f-0410-983b-f8d57d210c17