Adding initial Lydbær proposal and draft implementation for it in Max/MSP (which I've taken to calling "Maxbær").
The draft implementation currently exists of 3 objects:
   in≈ : this object takes multiple msp signals and acts as a source for the Lydbær graph
   op≈ : performs simple mathematical operations on a multichannel signal
   out≈ : provides multiple msp signals from a Lydbær graph, acting as the destination.

The in≈ object is not yet complete, the other two are theoretically complete.  None of them have been tested yet.  

The ≈ tag at the end of the object names was choosen because they process multiple channels as opposed to msp's ~ which processes a single channel.  The ≈ character on a U.S. keyboard (using a Mac) is option-x.  How to type this on a PC I am not sure.  I'm also not sure how to type the æ character on a PC.  It might be that we change these characters to something else if they don't seem to be a good choice.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@808 bc2ad3ca-9a4d-0410-8d14-35da9c29b421