Warning: Potentially destructive check-in.  The way all audio processing methods are implemented has changed, and all audio objects have been modified.  I have tested everything at least somewhat (on the Mac) and I'm not aware of any outstanding problems, but there is a possibility of something being broken.
All audio objects now implement their process methods as taking an *array* of input audio signals and an *array* of output audio signals.  Sidechains and sidechain process methods have been eliminated.  This change is required for allowing the creation of objects that can operate on any number of multichannel signals, for example to create an audio-mixing matrix object, and to generally support the Lydbær project.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@880 bc2ad3ca-9a4d-0410-8d14-35da9c29b421