Windows build fixes for Release targets.  The primary problem is that the _DEBUG symbol was defined for these release targets, and instead it should be NDEBUG.
Also, VisualStudio is crashing again on jcom.sum%.  It seems to be the % character that causes it to crash.  So the object has been renamed (at least on Windows) to jcom.jit.sum.mxe.  We can set up an object mapping to make the other name for it work, though I haven't done that at this time.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@4242 464221e2-0a0f-0410-983b-f8d57d210c17