Untested changes to the install script:
In theory, there should be no effect on the Mac installer.

In theory, this same script should now also build the installer for Windows (so that we don't have separate scripts in separate locations).  Furthermore, we needed to move from a simple shell script to Ruby so that we could do regular-expression substitutions on the Wix source files in order to get the TT extensions installing correctly.



git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@4262 464221e2-0a0f-0410-983b-f8d57d210c17