Irritating changes required to get TTBlue functioning properly on machines where Visual Studio is not installed, and you try to load an object in Max 5.0.5 (it was fine before in Max 5.0.4).
The most important thing is that I changed the options to make sure that Port Audio uses static-linking with the MS C Runtime, instead of dynamic.  Because this is different than the 'standard' portaudio DLL (which won't work) I renamed our version of it slightly.

There was also some wrangling to do with other project settings, including adding arcane prepocessor directives.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@889 bc2ad3ca-9a4d-0410-8d14-35da9c29b421