This change will break backwards compatibility with previously-compiled binaries (including Tap.Tools and old versions of Jamoma), and so everything that uses TTBlue will have to be recompiled (the size of the TTObject class has changed)...
We now maintained a 'valid' flag which indicates whether or not an object is in a state that is usable.  The reason for adding this flag is that there is so much work involved in setting up and tearing down a class that we have to protect against routines in the audio thread from being called when we are in the middle of freeing.

All of these added conditionals happening at every vector is certainly going to show up in performance.  The theory has been that this approach is cheaper than using "real" locks (e.g. critical regions) so long as it actually works.  As the system gets more complex this is called in question a bit and will need to be reviewed at some point.  We also need to see what the actual performance penalties are.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@899 bc2ad3ca-9a4d-0410-8d14-35da9c29b421