Brand new Xcode projects for every external in Jamoma.
These projects are largely configured through the use of a single .xcconfig file that controls shared settings for all of the projects.  We also do versioning differently which eliminates the need for the post-build perl script (which was both flaky and slow).

Finally, we do not link against the Max frameworks in /Library/Frameworks.  These frameworks being in this location is very problematic in some situations that we can discuss later.  We now link against them in a checkout of the Jamoma Toolbox (jatuuls project in GoogleCode).  More info about this will be posted to the jamoma-devel list shortly.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@4478 464221e2-0a0f-0410-983b-f8d57d210c17