Added a new target to build a static library ( libJamomaDSP.a ) on the Mac.
For Jamoma we will still use the compiled Mac framework, but Tap.Tools (starting in version 3.0.1) will use the static library.  This will should make it so that there is absolutely no tethering of Tap.Tools behavior and the version of Jamoma installed on machine.  

In Tap.Tools all usage of the DSP framework is encapsulated and private, so we can do this.  However, in a more distributed system (like Multicore) the static library should not and cannot be used because the symbol table instance must be shared among all objects -- so this will always use a dynamic library (such as the framework).


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@962 bc2ad3ca-9a4d-0410-8d14-35da9c29b421