Fix for, among other things, the broken behavior with /preset/read and other messages to the hub with no arguments (as reported by Mathieu and others).
There were multiple problems, but the one that everyone needs to remember is that you cannot send a message with no arguments to a jcom.message defined as msg_symbol -- a jcom.message defined as msg_symbol expects a symbol and requires it.  If the argument is optional then you need to use msg_generic.  

I'm not sure how we didn't catch the problem with these particular parameters sooner, but at least they are fixed now.


git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@4736 464221e2-0a0f-0410-983b-f8d57d210c17