Admitting defeat on the static lib version: there still ends up being trouble because when the extensions are loaded then they are linked to the dynamic version and the result is multiple copies of the symbol table which are not synchronized, memory begins leaking on (TTClass objects that are supposed to be stored but aren't because they are stored in another version of the lib).
The result is that we need to accept that the system is designed to be one shared framework that implements the environment for a machine.  This environment may be tapped by many clients, and may have many extensions, but there is just one framework.  It would be easier to change this restriction if we were coding in C, but by coding in C++ then trying to setup function pointer tables becomes a major chore.  

We need to embrace the one true framework principle, and then make sure we make the API as simple and stable as possible.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@969 bc2ad3ca-9a4d-0410-8d14-35da9c29b421