Some work on the trig_mapper. I implemented the features I added lately for the cont_mapper.  I was also trying to make the module a bit more user friendly (e.g. display of incoming values, disable/enable action options according to the chosen parameter type) I also added a "nOSC" trigger action - so that you can access a node with things like  ":/value/dec 2"
There is an inconsistency with the delay times: in cont_mapper the ramp-time is displayed in [ms]. In trig_mapper you have three different delaytimes - all in [s]. I changed the ramp-time to [ms] to be at least consistent with cont_mapper in this. But I am not sure about the repeat time and wait time. I think that has to be discussed on list.

git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/branches/active@5189 464221e2-0a0f-0410-983b-f8d57d210c17