New TTGraphics extension.
The TTGraphics extension encapsulates a light wrapper around Cairo ( http://cairographics.org/ ) and also provides a cross-platform class for opening a window onto which the Cairo wrapper can draw.

A simple demonstration of this is the new tt.graphics Max external.  This object simply opens a window and draws "Hello World!" in it.  While not changing the world, it shows how a class can create a new window and then respond to the window when it requests a re-draw.

While this code is being developed in a platform-independent manner, the actual tested implementation is Mac only for the time being.  Volunteers may freely work on the Windows side should the desire arise.


Now, a few notes about the implementation:

There is a TTGraphicsSurface class that is unused.  It may remain unused.  It is currently a remnant of one of the experiments I was doing trying to get things working.

There is a TTGraphicsWindow class.  This is where most of the action takes place.  This is likely to get a bit ugly because it is full of platform-specific code for creating the Window, getting Cairo surface, etc.  Because this class has the Window, this class also triggers re-draw notifications for the Window.  It then passes a the correct TTGraphicsContext to which drawing commands are sent.

The TTGraphicsContext is the light wrapper around Cairo.  You send messages to it when the context is valid (because the TTGraphicsWindow has made it valid) and they are composited to the Window.  Multiple objects may draw to the Window on any given cycle.  At the moment only a handful of functions from Cairo are actually wrapped by the TTGraphicsContext class; obviously this will be expanded.

To make this work all work, I had to compile the Cairo lib manually.  I noted what flags I used in the Wiki.


git-svn-id: http://ttblue.googlecode.com/svn/branches/active@979 bc2ad3ca-9a4d-0410-8d14-35da9c29b421