The dumb BlueButter plug-in now says "Hello World", even if I can't get it to display in the correct color.  This could be the worlds most labor-intensive 'Hello World' since the demise of punch-cards.
In the makeButter() method there is something going on with the bit ordering / composition causing the color problems that are plaguing me.  We should be provided with CAIRO_FORMAT_ARGB32, and that's fine.  But Apple's API seems dumb and doesn't want to play well with others.  

The desired output is a blue "Hello World" on a clear background.  When we set pixels here using char (variable c) then we get pink on blue, when we set the pixels here using unsigned int (variable d) then we get red on clear.  These results do not change even if you re-arrange the byte assignment ???   

After spending the whole day and night on this, I'm giving up for a little while.

git-svn-id: http://ttblue.googlecode.com/svn/branches/active@1030 bc2ad3ca-9a4d-0410-8d14-35da9c29b421