MulticoreOnRails: It is now actually working!  A lot of details follow:
There are some things about Rails I never knew before, and which are a major pain: namely that it is constantly reloading and re-initializing the app -- meaning that any instance variables in the controller are completely bogus from one call to the next, and so a Multicore chain gets rebuilt (usually incorrectly) *** EVERY TIME ANY REQUEST COMES INTO THE CONTROLLER *** (e.g. for every single tiny movement of the AJAX slider).

So we can get around this for the purposes of this example by using global variables instead.  I don't like this solution, but we're stuck with it for now.  In fact, with the default settings for the development environment, even this global code gets called upon every single action from the views!!!  I was able to stop that behavior by tweaking some of the rails config files.  Fortunately for me, I had been on the phone with Jesse Allison earlier tonight when he suggested that this development/reloading stuff could end up being a big problem for me!

There is still some more work to do get this Rails UI performing the way that it claims.  At the moment it is really just an oscillator, not an AM processor.  Things that need to happen to make the AM processor:

- we need to be able to connect to a second inlet on op≈
- the wavetable≈ object doesn't yet have a linearGain attr like it should
- yes, the gain slider has zipper noise -- there is already a FR for that in redmine