MulticoreOnRails: Moved the Graph creation out of the controllers and into the Rails App pre-initialization.  This achieves a few things: 1) Jesse was concerned about changing the caching settings for development as I did because of the requirement to constantly re-start the application, so those settings have been restored, 2) As Jesse explained it, the controllers are not the place for this kind of code because they are just for shuttling data around and it makes sense that they are created and destroyed as needed on-the-fly, 3) by moving to pre-initialization we are able to simply create a file that is only for the creating the graph ( in the config/preinitializers folder ) which makes the implementation really nice and clean.  A couple more notes:
The UI performance is definitely worse running in development as compared to deployment.  It is most noticeable when trying to drag the slider to control the master gain.

The type of file we are putting in the pre-initializers folder is the exact same format as we already export from Max and Pd!  So that makes it really easy to use those apps as editors for the graph used by the Rails app!