DSP library now compiled on the Mac as a dylib instead of a framework.
This is the first step toward support multiple independent Jamoma versions on the same machine.  The Clipper extension and jcom.clip~ external converted as well as the library, and used to prove that this is working -- which it is.  Now we just have to convert everything else in all of Jamoma (sigh).