Adding new function to functionLib: SmoothPolynomialFunction
This is a 5th order polynomial with the following properties:

	y = 6*pow(x, 5.) - 15*pow(x, 4.) + 10*pow(x, 3.0);
	y(0)   = 0
	y'(0)  = 0
	y''(0) = 0
	y(1)   = 1
    y'(1)  = 0
    y''(1) = 0

In other words: This function ramps from 0 to 1. The 1st and 2nd derivate equals 0 at x=0 and x=1, giving a smooth start and end to the ramp. The start and end of the ramp should be more smooth than for e.g. cosine, as the 2nd order derivative of cosine do not equal 0.