Starting refactoring of SpatLib with Jasch
Scene information will be embedded in the SpatClass, and we’ll be able to change renderers (spatialisation algorithms) dynamically. This implies that all of the following information will be stored in the SpatLib class:

- scene information that is global to all renderers and units
- scene information that is specific to one renderer only
- scene information that is global to all renderers, but specific to each source/sink entity
- scene information that is specific to one scene as well as each source/sink entity

For most or all renderers the scene will be over-specified, but one advantage of this is that we do not loose information about the scene when changing from renderer to another, we can reuse similar information in several renderers (e.g. source width), and when switching from render A to B and then back to A, none of the information relating to A has been lost in the process.

So fra we have refactored sources and sinks. Refactoring of renderers remains to be done.